+++ packages/moonstone/VirtualList/tests/VirtualList-specs.js
@@ -45,10 +45,10 @@ describe('VirtualList', () => {
 			resultScrollLeft = e.scrollLeft;
 			resultScrollTop = e.scrollTop;
 		};
-		renderItem = ({data, index, ...rest}) => {	// eslint-disable-line enact/display-name, enact/prop-types
+		renderItem = ({index, ...rest}) => {	// eslint-disable-line enact/display-name, enact/prop-types
 			return (
 				<Item {...rest}>
-					{data[index].name}
+					{items[index].name}
 				</Item>
 			);
 		};
@@ -79,7 +79,6 @@ describe('VirtualList', () => {
 		const subject = mount(
 			<VirtualList
 				clientSize={clientSize}
-				data={items}
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={30}
@@ -96,7 +95,6 @@ describe('VirtualList', () => {
 		const subject = mount(
 			<VirtualList
 				clientSize={clientSize}
-				data={items}
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={30}
@@ -113,7 +111,6 @@ describe('VirtualList', () => {
 		const subject = mount(
 			<VirtualList
 				clientSize={clientSize}
-				data={items}
 				dataSize={dataSize}
 				direction="horizontal"
 				itemRenderer={renderItem}
@@ -133,7 +130,6 @@ describe('VirtualList', () => {
 				<VirtualList
 					cbScrollTo={getScrollTo}
 					clientSize={clientSize}
-					data={items}
 					dataSize={dataSize}
 					itemRenderer={renderItem}
 					itemSize={30}
@@ -154,7 +150,6 @@ describe('VirtualList', () => {
 				<VirtualList
 					cbScrollTo={getScrollTo}
 					clientSize={clientSize}
-					data={items}
 					dataSize={dataSize}
 					direction="horizontal"
 					itemRenderer={renderItem}
@@ -176,7 +171,6 @@ describe('VirtualList', () => {
 				<VirtualList
 					cbScrollTo={getScrollTo}
 					clientSize={clientSize}
-					data={items}
 					dataSize={dataSize}
 					itemRenderer={renderItem}
 					itemSize={30}
@@ -198,7 +192,6 @@ describe('VirtualList', () => {
 					<VirtualList
 						cbScrollTo={getScrollTo}
 						clientSize={clientSize}
-						data={items}
 						dataSize={dataSize}
 						itemRenderer={renderItem}
 						itemSize={30}
@@ -219,7 +212,6 @@ describe('VirtualList', () => {
 					<VirtualList
 						cbScrollTo={getScrollTo}
 						clientSize={clientSize}
-						data={items}
 						dataSize={dataSize}
 						itemRenderer={renderItem}
 						itemSize={30}
@@ -240,7 +232,6 @@ describe('VirtualList', () => {
 					<VirtualList
 						cbScrollTo={getScrollTo}
 						clientSize={clientSize}
-						data={items}
 						dataSize={dataSize}
 						itemRenderer={renderItem}
 						itemSize={30}
@@ -261,11 +252,17 @@ describe('VirtualList', () => {
 	describe('Adding an item', () => {
 		it('should render an added item named \'Password 0\' as the first item', (done) => {
 			const itemArray = [{name: 'A'}, {name: 'B'}, {name: 'C'}];
+			const renderItem = ({index, ...rest}) => { // eslint-disable-line enact/display-name, enact/prop-types
+				return (
+					<div {...rest} id={'item' + index}>
+						{itemArray[index].name}
+					</div>
+				);
+			};
 
 			const subject = mount(
 				<VirtualList
 					clientSize={clientSize}
-					data={itemArray}
 					dataSize={itemArray.length}
 					itemRenderer={renderItem}
 					itemSize={30}
@@ -273,7 +270,7 @@ describe('VirtualList', () => {
 			);
 
 			itemArray.unshift({name: 'Password 0'});
-			subject.setProps({data: itemArray, dataSize: itemArray.length});
+			subject.setProps({dataSize: itemArray.length});
 
 			setTimeout(() => {
 				const expected = itemArray[0].name;