+++ packages/sampler/stories/moonstone-stories/VirtualList.js
@@ -15,8 +15,9 @@ const Config = mergeComponentMetadata('VirtualList', VirtualList, VirtualListBas
 
 const
 	items = [],
+	isDisabledItem = (index) => (items[index].disabled),
 	// eslint-disable-next-line enact/prop-types, enact/display-name
-	renderItem = (size) => ({data, index, ...rest}) => {
+	renderItem = (size) => ({index, ...rest}) => {
 		const itemStyle = {
 			height: size + 'px',
 			borderBottom: ri.unit(3, 'rem') + ' solid #202328',
@@ -25,7 +26,7 @@ const
 
 		return (
 			<Item {...rest} style={itemStyle}>
-				{data[index]}
+				{items[index]}
 			</Item>
 		);
 	};
@@ -44,7 +45,6 @@ storiesOf('UI', module)
 			const itemSize = ri.scale(number('itemSize', 72));
 			return (
 				<UiVirtualList
-					data={items}
 					dataSize={number('dataSize', items.length)}
 					itemRenderer={renderItem(itemSize)}
 					itemSize={itemSize}
@@ -69,9 +69,9 @@ storiesOf('Moonstone', module)
 			const itemSize = ri.scale(number('itemSize', 72));
 			return (
 				<VirtualList
-					data={items}
 					dataSize={number('dataSize', items.length)}
 					focusableScrollbar={nullify(boolean('focusableScrollbar', false))}
+					isDisabledItem={isDisabledItem}
 					itemRenderer={renderItem(itemSize)}
 					itemSize={itemSize}
 					onScrollStart={action('onScrollStart')}