+++ packages/ui/VirtualList/tests/VirtualList-specs.js
@@ -44,10 +44,10 @@ describe('VirtualList', () => {
 			resultScrollLeft = e.scrollLeft;
 			resultScrollTop = e.scrollTop;
 		};
-		renderItem = ({data, index, ...rest}) => {	// eslint-disable-line enact/display-name, enact/prop-types
+		renderItem = ({index, ...rest}) => {	// eslint-disable-line enact/display-name, enact/prop-types
 			return (
 				<div {...rest} id={'item' + index}>
-					{data[index].name}
+					{items[index].name}
 				</div>
 			);
 		};
@@ -78,7 +78,6 @@ describe('VirtualList', () => {
 		const subject = mount(
 			<VirtualList
 				clientSize={clientSize}
-				data={items}
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={30}
@@ -95,7 +94,6 @@ describe('VirtualList', () => {
 		const subject = mount(
 			<VirtualList
 				clientSize={clientSize}
-				data={items}
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={30}
@@ -114,7 +112,6 @@ describe('VirtualList', () => {
 				<VirtualList
 					cbScrollTo={getScrollTo}
 					clientSize={clientSize}
-					data={items}
 					dataSize={dataSize}
 					itemRenderer={renderItem}
 					itemSize={30}
@@ -135,7 +132,6 @@ describe('VirtualList', () => {
 				<VirtualList
 					cbScrollTo={getScrollTo}
 					clientSize={clientSize}
-					data={items}
 					dataSize={dataSize}
 					direction="horizontal"
 					itemRenderer={renderItem}
@@ -157,7 +153,6 @@ describe('VirtualList', () => {
 				<VirtualList
 					cbScrollTo={getScrollTo}
 					clientSize={clientSize}
-					data={items}
 					dataSize={dataSize}
 					itemRenderer={renderItem}
 					itemSize={30}
@@ -179,7 +174,6 @@ describe('VirtualList', () => {
 					<VirtualList
 						cbScrollTo={getScrollTo}
 						clientSize={clientSize}
-						data={items}
 						dataSize={dataSize}
 						itemRenderer={renderItem}
 						itemSize={30}
@@ -200,7 +194,6 @@ describe('VirtualList', () => {
 					<VirtualList
 						cbScrollTo={getScrollTo}
 						clientSize={clientSize}
-						data={items}
 						dataSize={dataSize}
 						itemRenderer={renderItem}
 						itemSize={30}
@@ -221,7 +214,6 @@ describe('VirtualList', () => {
 					<VirtualList
 						cbScrollTo={getScrollTo}
 						clientSize={clientSize}
-						data={items}
 						dataSize={dataSize}
 						itemRenderer={renderItem}
 						itemSize={30}
@@ -242,11 +234,17 @@ describe('VirtualList', () => {
 	describe('Adding an item', () => {
 		it('should render an added item named \'Password 0\' as the first item', (done) => {
 			const itemArray = [{name: 'A'}, {name: 'B'}, {name: 'C'}];
+			const renderItem = ({index, ...rest}) => { // eslint-disable-line enact/display-name, enact/prop-types
+				return (
+					<div {...rest} id={'item' + index}>
+						{itemArray[index].name}
+					</div>
+				);
+			};
 
 			const subject = mount(
 				<VirtualList
 					clientSize={clientSize}
-					data={itemArray}
 					dataSize={itemArray.length}
 					itemRenderer={renderItem}
 					itemSize={30}
@@ -254,7 +252,7 @@ describe('VirtualList', () => {
 			);
 
 			itemArray.unshift({name: 'Password 0'});
-			subject.setProps({data: itemArray, dataSize: itemArray.length});
+			subject.setProps({dataSize: itemArray.length});
 
 			setTimeout(() => {
 				const expected = itemArray[0].name;